convert tiller_init to general helm_env script

+++ terraform/google/scripts/helm_env.sh
@@ -8,6 +8,7 @@ if [ ! -f "${cert_dir}/ca.key.pem.asc" ];then
 	openssl genrsa -out "${cert_dir}/ca.key.pem" 4096
 	gpg -ear "${gpg_key}" "${cert_dir}/ca.key.pem"
 fi
+
 if [ ! -f "${cert_dir}/ca.cert.pem" ];then
 	openssl req \
 		-new -x509 \
@@ -19,10 +20,11 @@ if [ ! -f "${cert_dir}/ca.cert.pem" ];then
 		-subj "${cert_subject}"
 fi
 
-if [ ! -f "${cert_dir}/tiller.key.pem" ];then
+if [ ! -f "${cert_dir}/tiller.key.pem.asc" ];then
 	openssl genrsa -out "${cert_dir}/tiller.key.pem" 4096
 	gpg -ear "${gpg_key}" "${cert_dir}/tiller.key.pem"
 fi
+
 if [ ! -f "${cert_dir}/tiller.csr.pem" ];then
 	openssl req \
 		-new \
@@ -31,6 +33,7 @@ if [ ! -f "${cert_dir}/tiller.csr.pem" ];then
 		-key "${cert_dir}/tiller.key.pem" \
 		-out "${cert_dir}/tiller.csr.pem"
 fi
+
 if [ ! -f "${cert_dir}/tiller.cert.pem" ];then
 	openssl x509 \
 		-req \
@@ -42,10 +45,11 @@ if [ ! -f "${cert_dir}/tiller.cert.pem" ];then
 		-out "${cert_dir}/tiller.cert.pem"
 fi
 
-if [ ! -f "${cert_dir}/helm.key.pem" ];then
+if [ ! -f "${cert_dir}/helm.key.pem.asc" ];then
 	openssl genrsa -out "${cert_dir}/helm.key.pem" 4096
 	gpg -ear "${gpg_key}" "${cert_dir}/helm.key.pem"
 fi
+
 if [ ! -f "${cert_dir}/helm.csr.pem" ];then
 	openssl req \
 		-new \
@@ -54,6 +58,7 @@ if [ ! -f "${cert_dir}/helm.csr.pem" ];then
 		-key "${cert_dir}/helm.key.pem" \
 		-out "${cert_dir}/helm.csr.pem"
 fi
+
 if [ ! -f "${cert_dir}/helm.cert.pem" ];then
 	openssl x509 \
 		-req \
@@ -65,18 +70,24 @@ if [ ! -f "${cert_dir}/helm.cert.pem" ];then
 		-out "${cert_dir}/helm.cert.pem"
 fi
 
+if [ -f "${cert_dir}/tiller.key.pem" ]; then
+	helm init \
+		--service-account tiller \
+		--upgrade \
+		--tiller-tls \
+		--tiller-tls-verify \
+		--tls-ca-cert "${cert_dir}/ca.cert.pem" \
+		--tiller-tls-cert "${cert_dir}/tiller.cert.pem" \
+		--tiller-tls-key "${cert_dir}/tiller.key.pem"
+	rm "${cert_dir}/ca.key.pem"
+	rm "${cert_dir}/tiller.key.pem"
+	rm "${cert_dir}/helm.key.pem"
+fi
 
 source scripts/kube_env.sh
 
-helm init \
-	--service-account tiller \
-	--upgrade \
-	--tiller-tls \
-	--tiller-tls-verify \
-	--tls-ca-cert "${cert_dir}/ca.cert.pem" \
-	--tiller-tls-cert "${cert_dir}/tiller.cert.pem" \
-	--tiller-tls-key "${cert_dir}/tiller.key.pem"
-
-rm "${cert_dir}/ca.key.pem"
-rm "${cert_dir}/tiller.key.pem"
-rm "${cert_dir}/helm.key.pem"
+HELM_HOME="$(mktemp -d -p /dev/shm/)"
+cp "${cert_dir}/ca.cert.pem" "${HELM_HOME}/"
+cp "${cert_dir}/helm.cert.pem" "${HELM_HOME}/"
+gpg -d "${cert_dir}/helm.key.pem.asc" > "${HELM_HOME}/helm.key.pem"
+export HELM_HOME
